+++ tests/test_model_v4.py
@@ -3,10 +3,11 @@ import datetime
 import geojson
 import pytest
 
+from pyodata.policies import PolicyIgnore
 from pyodata.model.builder import MetadataBuilder
-from pyodata.exceptions import PyODataModelError
+from pyodata.exceptions import PyODataModelError, PyODataException, PyODataParserError
 from pyodata.model.type_traits import TypTraits
-from pyodata.model.elements import Types
+from pyodata.model.elements import Types, TypeInfo, NullType
 
 from pyodata.config import Config
 from pyodata.v4 import ODataV4
@@ -172,3 +173,93 @@ def test_referential_constraint(schema_v4):
     assert repr(nav_property.referential_constraints[0]) == \
         'ReferentialConstraint(StructTypeProperty(CategoryID), StructTypeProperty(ID))'
 
+
+def test_enum_parsing(schema_v4):
+    """Test correct parsing of enum"""
+
+    country = schema_v4.enum_type('Country').USA
+    assert str(country) == "Country'USA'"
+
+    country2 = schema_v4.enum_type('Country')['USA']
+    assert str(country2) == "Country'USA'"
+
+    try:
+        schema_v4.enum_type('Country').Cyprus
+    except PyODataException as ex:
+        assert str(ex) == f'EnumType EnumType(Country) has no member Cyprus'
+
+    c = schema_v4.enum_type('Country')[1]
+    assert str(c) == "Country'China'"
+
+    try:
+        schema_v4.enum_type('Country')[15]
+    except PyODataException as ex:
+        assert str(ex) == f'EnumType EnumType(Country) has no member with value {15}'
+
+    type_info = TypeInfo(namespace=None, name='Country', is_collection=False)
+
+    try:
+        schema_v4.get_type(type_info)
+    except PyODataModelError as ex:
+        assert str(ex) == f'Neither primitive types nor types parsed from service metadata contain requested type {type_info[0]}'
+
+    language = schema_v4.enum_type('Language')
+    assert language.is_flags is True
+
+    try:
+        schema_v4.enum_type('ThisEnumDoesNotExist')
+    except KeyError as ex:
+        assert str(ex) == f'\'EnumType ThisEnumDoesNotExist does not exist in any Schema Namespace\''
+
+    try:
+        schema_v4.enum_type('Country', 'WrongNamespace').USA
+    except KeyError as ex:
+        assert str(ex) == '\'EnumType Country does not exist in Schema Namespace WrongNamespace\''
+
+
+def test_unsupported_enum_underlying_type(xml_builder_factory):
+    """Test if parser will parse only allowed underlying types"""
+    xml_builder = xml_builder_factory()
+    xml_builder.add_schema('Test', '<EnumType Name="UnsupportedEnumType" UnderlyingType="Edm.Bool" />')
+    xml = xml_builder.serialize()
+
+    try:
+        MetadataBuilder(xml, Config(ODataV4)).build()
+    except PyODataParserError as ex:
+        assert str(ex).startswith(f'Type Edm.Bool is not valid as underlying type for EnumType - must be one of')
+
+
+def test_enum_value_out_of_range(xml_builder_factory):
+    """Test if parser will check for values ot of range defined by underlying type"""
+    xml_builder = xml_builder_factory()
+    xml_builder.add_schema('Test', """
+        <EnumType Name="Num" UnderlyingType="Edm.Byte">
+            <Member Name="TooBig" Value="-130" />
+        </EnumType>
+        """)
+    xml = xml_builder.serialize()
+
+    try:
+        MetadataBuilder(xml, Config(ODataV4)).build()
+    except BaseException as ex:
+        assert str(ex) == f'Value -130 is out of range for type Edm.Byte'
+
+
+def test_enum_null_type(xml_builder_factory):
+    """ Test NullType being correctly assigned to invalid types"""
+    xml_builder = xml_builder_factory()
+    xml_builder.add_schema('TEST.NAMESPACE', """
+        <EnumType Name="MasterEnum" UnderlyingType="Edm.String" />
+    """)
+
+    metadata = MetadataBuilder(
+        xml_builder.serialize(),
+        config=Config(
+            ODataV4,
+            default_error_policy=PolicyIgnore()
+        ))
+
+    schema = metadata.build()
+
+    type_info = TypeInfo(namespace=None, name='MasterEnum', is_collection=False)
+    assert isinstance(schema.get_type(type_info), NullType)
